+++ Tests/EZNetworkingTests/Mocks/MockURLSession.swift
@@ -72,37 +72,3 @@ extension MockURLSession {
         }
     }
 }
-
-class MockURLSessionDataTask: URLSessionDataTask {
-    private let closure: () -> Void
-    var didCancel: Bool = false
-    
-    init(closure: @escaping () -> Void) {
-        self.closure = closure
-    }
-    
-    override func resume() {
-        closure()
-    }
-    
-    override func cancel() {
-        didCancel = true
-    }
-}
-
-class MockURLSessionDownloadTask: URLSessionDownloadTask {
-    private let closure: () -> Void
-    var didCancel: Bool = false
-
-    init(closure: @escaping () -> Void) {
-        self.closure = closure
-    }
-    
-    override func resume() {
-        closure()
-    }
-    
-    override func cancel() {
-        didCancel = true
-    }
-}
