+++ Tests/EZNetworkingTests/Mocks/MockURLSessionDownloadTask.swift
@@ -1,0 +1,19 @@
+import Foundation
+import EZNetworking
+
+class MockURLSessionDownloadTask: URLSessionDownloadTask {
+    private let closure: () -> Void
+    var didCancel: Bool = false
+
+    init(closure: @escaping () -> Void) {
+        self.closure = closure
+    }
+    
+    override func resume() {
+        closure()
+    }
+    
+    override func cancel() {
+        didCancel = true
+    }
+}
